server/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
PyMySQL==1.1.0
cryptography==41.0.8

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Processing
Pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0

# Text-to-Speech
pyttsx3==2.90
gTTS==2.4.0

# Braille Conversion
louis==1.3

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6

# File Upload
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2
